test: consolidate migration manager tests into single file
